Add dummy line object removal
When a dummy legend entry is appended a Chart Line Object containing a
single NaN entry is added to the axes object. With the previous legend
entry removal methodology these objects would not be removed and could
accumulate in the axes without the user knowing. These objects are now
removed when their corresponding legend entry is removed.