Improve detection of dummy entries
Detection was based on the number of elements of a lineseries object and
the value being NaN. This implicitly assumes only to be the case for dummy
lineseries created by legtools, which is false. By using the UserData
property when the dummy is created, the probability of a lineseries object
being a legtools dummy is practically equal to one.