Fix bug in append method
The append method contains a bug when it appends to a legend that has
been permuted previously or if the parent axes children have been
sorted differently after the legend has been modified. Append assumes
the parent axes children are in the same order as the current legend
PlotChildren, which may not be the case. For example, after permuting
the legend entries, this order may have changed.