Add index magnitude check for entry removal
Previous versions of the remove method left index magnitude error
checking up to MATLAB, which works fine except for the case where only
one index is provided to a legend with one entry. Because the check was
only a magnitude one, the legend handle is removed in this case even if
the index is not 1. A check of the indices' magnitude has been added to
mitigate this. The user will be warned if indices have been thrown out.

Resolves: #8