Enforce adddummy plot parameter input format
For multiple dummy entries, adddummy expects a cell array of cells,
where each top level cell corresponds to a new legend string. To ensure
consistent behavior without convoluted checks, the character array plot
parameter input is double wrapped into a cell array. This same
enforcement is not present for the case of a single level cell array so
the expansion attempt by the plot call errors out. A check for this case
has been added and the plot parameters are nested when necessary.