Add order consideration for appending
The existing append method assumes that the order of the existing legend
entries has not been permuted from the plot order of the axes' Children,
and vice-versa. Because of this assumption, the order of the legend
strings can diverge from the axes' Children plot order for reasons that
are not clear to the user. A union of the two object arrays is taken in
order to account for the permutation potential.

A stable union of the legend's PlotChildren and the parent Axes'
Children will return the object handles existing in the legend, in the
order they appear in the legend, followed by the object handles not in
the legend, in the order they are rendered in the parent Axes.

Add warning when more strings are passed to append than there are legend
entries. This warning is consistent with the behavior of legend.

See: #11