Restore append functionality for MATLAB >= R2017a
The addition of 'AutoUpdate' to legend in R2017a breaks the
functionality of append. With 'AutoUpdate' turned off we can restore
the functionality of legtools, but turning it back on causes our
appended legend entries to be deleted. Clearing out the undocumented
'PlotChildrenExcluded' legend property seems to prevent this from
occuring

This is untested in versions < R2017a

See #15